+++ packages/docs/site/CHANGELOG.md
@@ -3,6 +3,11 @@
 - reinstated `/ui` docs
 - reorganised `/ui` examples
 - /components: made the render area scrollable
+- support documenting components with unnamed slots
+	- caveats:
+		- as of now dynamic slots aren't supported so we can't document named slots
+		- it supports 1 level of depth only so slots being components with slots
+			themselves won't work
 
 ## `@svizzle/site` v0.4.0